Fix ServerMessage deserialize priority
Server messages with an error and an id attached were getting decoded to
ServerMessage::Error instead of ServerMessage::Response (losing the id). This
was because serde matched the former shape before trying to match the latter,
and with unknown keys allowed it fit successfully. Changing the order here fixes
this.