Cow, rustfmt, baking, and license handling
Cow: CDP protocol types now use Cow<'a, str> instead of String. The 'a lifetime parameter is
propagated upward to the types that need it. This allows for more efficient serialization, allowing
the use of borrowed strings instead of potentially forcing a copy.

rustfmt: To prevent build error messages that fill the console with a single gigantic line of text,
the generated source is now run through rustfmt. But I'd rather not force end users to depend in
any way on rustfmt (even just via build-dependencies), so...

Baking: Releases of the cdp crate shall have the generated source "baked" into the published
artifact. A script for performing this task, `bake.sh`, has been added.

License handling: I've clarified the license status vis-a-vis the BSD 3-Clause license which
applies to the source JSON specification files.